requirements.txt
# ============================================================
# Dépendances — Robot Éligibilité Énergétique
# pip install -r requirements.txt
# ============================================================

# Serveur web
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Twilio (appels téléphoniques)
twilio==9.3.6

# OpenAI (GPT-4o + Whisper)
openai==1.54.4

# Google Cloud TTS Neural2 (synthèse vocale)
google-cloud-texttospeech==2.17.2
google-auth==2.36.0

# Excel
openpyxl==3.1.5

# HTTP asynchrone (caller.py)
httpx==0.27.2

# Variables d'environnement
python-dotenv==1.0.1

# Parsing form data (requis par FastAPI pour les webhooks Twilio)
python-multipart==0.0.12
